backend/requirements.txt
# API
fastapi
uvicorn
pydantic
python-dotenv
loguru

# Data
pandas
numpy<2.0
scikit-learn
tqdm
requests

# RAG
sentence-transformers
faiss-cpu

# LLM
google-generativeai==0.3.2
transformers
sentencepiece
